yocto: update software packages to openEuler-22.03-LTS-SP3
* Update commit IDs to the latest commit ID in branch
  openEuler-22.03-LTS-SP3 of each software package repository. If the
  repository does not contain the specific branch, we should analyze it
  manually.

* upgrade graphicsmagick from 1.3.38 to 1.3.41

* downgrade zlib from 1.2.13 to 1.2.11

* downgrade python3-pep8 from 1.7.1 to 1.5.7

* upgrade pyyaml from 5.3.1 to 5.4.1

the latest commit of PyYAML in branch openEuler-22.03-LTS-SP3 has
version 6.0. However, after comparing to the bb files in upstream
community, we can find that compiling PyYAML 6.0 requires Yocto 4.0+
version. The latest version Yocto 3.3.x can support is PyYAML 5.4.1.

* upgrade xkeyboard-config from 2.33 to 2.38

Since the version 2.38 uses meson for building instead of original "make",
we have to add the bb file of version 2.38 into our repository.

* upgrade kernel version from 5.10.0-153.12.0 to 5.10.0-161.0.0

The kernel version in branch openEuler-22.03-LTS-SP3 of repository
"src-openEuler/kernel" is 5.10.0-171.0.0. Nevertheless, the latest
version of kernel which can be applied to the preempt-rt patch is
5.10.0-161.0.0.

* the yocto-embedded-tools repository remains unchanged

It is because when building our OS, we only depend on the software
packages related to dsoftbus and ros, which remain unchanged for last
half of a year.

* the openAMP and libmetal remains unchanged

Since our openAMP based MICA framework relies on the 2022.10 version of
openAMP

* xerces-c remains unchanged

Because only in version 3.2.4+ can the specific building error be
resolved

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>