ca-certificates: fix bug for containing empty CA list
* remove useless pkg_preinst part
* run update-ca-trust script on target
* add p11-kit package

ca-certificates generates PKCS#11 module containing all the certificates
of trusted root CA, and the update-ca-trust script calls p11-kit to
extract certificates from the PKCS#11 module.

Since we build ca-certificates from scratch, it is no need for us to
save certificates first, so remove the unnecessary part of code.

* p11-kit uses dynamically built meson-native package rather than the
  one in nativesdk with version 0.61.5

p11-kit requires a higher version of meson than the one provided by
nativesdk, which has the version 0.61.5. However, if we use a higher
version of meson to provide native package, it will break the build of
systemd.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>