docs: add intro to openeuler-container-os
container os is a tiny image for running containers based on it.
openeuler-container-os holds the openEuler 23.09 server version
OS container image, and will attach to it each time root user
logs in. Show user how to build and how to use it through words
and a demo video.

add more introduction to isula commands.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>