manifest: refactor to fix do fetch error
* reorder the software packages in alphabetical order

* add .git suffix to all git repository.

Some packages in SP3 have url with .git suffix, like diffutils, while they
have url without .git suffix in master branch. If developers switch branch
from SP3 to master to build a new OS image, it will cause an error when
do_fetch tasks are executed. To prevent this issue, we add ".git" suffix
to all urls, which makes sense because it is the correct way to represent
a Git repository.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>