kubeedge.cfg: add more options so that it's a more complete set
kubeedge.cfg is a set of kernel configs for enabling running
kubeedge on openEuler Embedded, add more options so that the
hieulerpi1 can be enable this feature by add this scc.

However, this is still not a complete set, given that scc is a
file for appending kernel configs to existing config file.
Different config file may need different set of configs to enable
kubeedge feature.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>